# gen/openapiv2/ride_platform.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "ride/analytics/v1/analytics.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/v1/me/wallet-pin": {
      "get": {
        "summary": "GetWalletPinStatus says whether the caller has a PIN and whether it is\nlocked.",
        "operationId": "WalletPinService_GetWalletPinStatus",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1WalletPinStatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "WalletPinService"
        ]
      },
      "put": {
        "summary": "SetWalletPin sets the caller's PIN: 4 or 6 digits, not all the same and\nnot a straight run (1234, 654321). The first time, just new_pin. To\nchange it, current_pin too (a wrong one counts as a wrong attempt). A\nforgotten PIN: sign in again (with a one-time code), then set a new one\nwithin 10 minutes of signing in without current_pin; this also lifts a\nlock.",
        "operationId": "WalletPinService_SetWalletPin",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1WalletPinStatusResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1SetWalletPinRequest"
            }
          }
        ],
        "tags": [
          "WalletPinService"
        ]
      }
    },
...
  },
  "definitions": {
...
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
...
    "v1SetWalletPinRequest": {
      "type": "object",
      "properties": {
        "newPin": {
          "type": "string"
        },
        "currentPin": {
          "type": "string"
        }
      }
    },
...
    "v1WalletPinStatusResponse": {
      "type": "object",
      "properties": {
        "isSet": {
          "type": "boolean"
        },
        "setAt": {
          "type": "string",
          "format": "date-time"
        },
        "lockedUntil": {
          "type": "string",
          "format": "date-time",
          "description": "Set while wrong PINs keep it locked."
        },
        "attemptsLeft": {
          "type": "integer",
          "format": "int32",
          "description": "Wrong PINs left before it locks."
        }
      }
    },
...
  }
}